# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/Medicine": {
            "get": {
                "description": "list Medicine entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List Medicine entities",
                "operationId": "list-Medicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Medicine"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Medicine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Medicine",
                "operationId": "create-Medicine",
                "parameters": [
                    {
                        "description": "Medicine entity",
                        "name": "Medicine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Medicine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Medicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/UnitOfMedicine": {
            "get": {
                "description": "list UnitOfMedicine entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List UnitOfMedicine entities",
                "operationId": "list-UnitOfMedicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.UnitOfMedicine"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create UnitOfMedicine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create UnitOfMedicine",
                "operationId": "create-UnitOfMedicine",
                "parameters": [
                    {
                        "description": "UnitOfMedicine entity",
                        "name": "UnitOfMedicine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.UnitOfMedicine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.UnitOfMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/UnitOfMedicine/{id}": {
            "get": {
                "description": "get UnitOfMedicine by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a UnitOfMedicine entity by ID",
                "operationId": "get-UnitOfMedicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UnitOfMedicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.UnitOfMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update UnitOfMedicine by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a UnitOfMedicine entity by ID",
                "operationId": "update-UnitOfMedicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UnitOfMedicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UnitOfMedicine entity",
                        "name": "UnitOfMedicine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.UnitOfMedicine"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.UnitOfMedicine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get UnitOfMedicine by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a UnitOfMedicine entity by ID",
                "operationId": "delete-UnitOfMedicine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UnitOfMedicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Medicine": {
            "type": "object",
            "properties": {
                "levelOfDangerousID": {
                    "type": "integer"
                },
                "medicineTypeID": {
                    "type": "integer"
                },
                "unitOfMedicineID": {
                    "type": "integer"
                }
            }
        },
...
        "ent.DispenseMedicine": {
            "type": "object",
            "properties": {
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DispenseMedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DispenseMedicineEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.DrugAllergy": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "DateTime holds the value of the \"dateTime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DrugAllergyQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DrugAllergyEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DrugAllergyEdges": {
            "type": "object",
            "properties": {
                "medicine": {
                    "description": "Medicine holds the value of the medicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Medicine"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientInfo"
                },
                "pharmacist": {
                    "description": "Pharmacist holds the value of the pharmacist edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Pharmacist"
                }
            }
        },
        "ent.LevelOfDangerous": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the LevelOfDangerousQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.LevelOfDangerousEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.LevelOfDangerousEdges": {
            "type": "object",
            "properties": {
                "medicine": {
                    "description": "Medicine holds the value of the Medicine edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Medicine"
                    }
                }
            }
        },
        "ent.Medicine": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "integer"
                },
                "brand": {
                    "description": "Brand holds the value of the \"brand\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineEdges"
                },
                "howtouse": {
                    "description": "Howtouse holds the value of the \"howtouse\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "price": {
                    "description": "Price holds the value of the \"price\" field.",
                    "type": "number"
                },
                "serial": {
                    "description": "Serial holds the value of the \"serial\" field.",
                    "type": "string"
                }
            }
        },
        "ent.MedicineEdges": {
            "type": "object",
            "properties": {
                "drugallergys": {
                    "description": "Drugallergys holds the value of the drugallergys edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.DrugAllergy"
                    }
                },
                "levelOfDangerous": {
                    "description": "LevelOfDangerous holds the value of the LevelOfDangerous edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.LevelOfDangerous"
                },
                "medicineType": {
                    "description": "MedicineType holds the value of the MedicineType edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineType"
                },
                "medicinepresciption": {
                    "description": "Medicinepresciption holds the value of the medicinepresciption edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                },
                "order": {
                    "description": "Order holds the value of the order edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Order"
                    }
                },
                "unitOfMedicine": {
                    "description": "UnitOfMedicine holds the value of the UnitOfMedicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UnitOfMedicine"
                }
            }
        },
        "ent.MedicineType": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicineTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.MedicineTypeEdges": {
            "type": "object",
            "properties": {
                "medicine": {
                    "description": "Medicine holds the value of the Medicine edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Medicine"
                    }
                }
            }
        },
        "ent.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "integer"
                },
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OrderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OrderEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "price": {
                    "description": "Price holds the value of the \"price\" field.",
                    "type": "number"
                },
                "total": {
                    "description": "Total holds the value of the \"total\" field.",
                    "type": "number"
                }
            }
        },
        "ent.OrderEdges": {
            "type": "object",
            "properties": {
                "company": {
                    "description": "Company holds the value of the company edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Company"
                    }
                },
                "medicine": {
                    "description": "Medicine holds the value of the medicine edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Medicine"
                    }
                },
                "pharmacist": {
                    "description": "Pharmacist holds the value of the pharmacist edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Pharmacist"
                }
            }
        },
        "ent.PatientInfo": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "Age holds the value of the \"age\" field.",
                    "type": "integer"
                },
                "cardNumber": {
                    "description": "CardNumber holds the value of the \"cardNumber\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientInfoQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientInfoEdges"
                },
                "gender": {
                    "description": "Gender holds the value of the \"gender\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Pharmacist": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PharmacistQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PharmacistEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Prescription": {
            "type": "object",
            "properties": {
                "Value": {
                    "description": "Value holds the value of the \"Value\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrescriptionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrescriptionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PrescriptionEdges": {
            "type": "object",
            "properties": {
                "dispensemedicine": {
                    "description": "Dispensemedicine holds the value of the dispensemedicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DispenseMedicine"
                },
                "prescriptiondoctor": {
                    "description": "Prescriptiondoctor holds the value of the prescriptiondoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "prescriptionmedicine": {
                    "description": "Prescriptionmedicine holds the value of the prescriptionmedicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Medicine"
                },
                "prescriptionpatient": {
                    "description": "Prescriptionpatient holds the value of the prescriptionpatient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientInfo"
                }
            }
        },
        "ent.UnitOfMedicine": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UnitOfMedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UnitOfMedicineEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
        "ent.UnitOfMedicineEdges": {
            "type": "object",
            "properties": {
                "medicine": {
                    "description": "Medicine holds the value of the Medicine edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Medicine"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}